requirements.txt
numpy
scipy
matplotlib
numexpr
cython
django-celery==3.1.10
traits
pandas
patsy
statsmodels
PyOpenGL 
PyOpenGL_accelerate
Django==1.6
pylibftdi
nitime
sphinx
numpydoc
tornado
tables
pyserial
h5py
pygame
south